pyproject.toml
[project]
name = "rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.124.2",
    "google>=3.0.0",
    "google-genai>=1.55.0",
    "inngest>=0.5.13",
    "llama-index-core>=0.14.10",
    "llama-index-readers-file>=0.5.5",
    "openai>=2.9.0",
    "python-dotenv>=1.2.1",
    "qdrant-client==1.15.1",
    "streamlit>=1.52.1",
    "uvicorn>=0.38.0",
]
